+++ mh/medici/couchbase/couchbase-service.properties
@@ -28,14 +28,18 @@ clusterBucketType ="couchbase"
 
 clusterReplicatCount=1
 
+cbVersion="2.0.0"
 /* Taken from http://www.couchbase.com/download */
-couchbaseRpmLinux32="http://packages.couchbase.com/releases/2.0.0/couchbase-server-community_x86_2.0.0.rpm"
-couchbaseRpmLinux64="http://packages.couchbase.com/releases/2.0.0/couchbase-server-community_x86_64_2.0.0.rpm"
+couchbaseRpmLinux32="http://packages.couchbase.com/releases/${cbVersion}/couchbase-server-community_x86_${cbVersion}.rpm"
+couchbaseRpmLinux64="http://packages.couchbase.com/releases/${cbVersion}/couchbase-server-community_x86_64_${cbVersion}.rpm"
 
-couchbaseDebUbuntu32="http://packages.couchbase.com/releases/2.0.0/couchbase-server-community_x86_2.0.0.deb"
-couchbaseDebUbuntu64="http://packages.couchbase.com/releases/2.0.0/couchbase-server-community_x86_64_2.0.0.deb"
+couchbaseDebUbuntu32="http://packages.couchbase.com/releases/${cbVersion}/couchbase-server-community_x86_${cbVersion}.deb"
+couchbaseDebUbuntu64="http://packages.couchbase.com/releases/${cbVersion}/couchbase-server-community_x86_64_${cbVersion}.deb"
 
-couchbaseExeWindows32="http://packages.couchbase.com/releases/2.0.0/couchbase-server-community_x86_2.0.0.setup.exe"
-couchbaseExeWindows64="http://packages.couchbase.com/releases/2.0.0/couchbase-server-community_x86_64_2.0.0.setup.exe"
+couchbaseExeWindows32="http://packages.couchbase.com/releases/${cbVersion}/couchbase-server-community_x86_${cbVersion}.setup.exe"
+couchbaseExeWindows64="http://packages.couchbase.com/releases/${cbVersion}/couchbase-server-community_x86_64_${cbVersion}.setup.exe"
 
-couchbaseZipMac64="http://packages.couchbase.com/releases/2.0.0/couchbase-server-community_x86_64_2.0.0.zip"
+couchbaseZipMac64="http://packages.couchbase.com/releases/${cbVersion}/couchbase-server-community_x86_64_${cbVersion}.zip"
+
+/* Only for a demo */
+dataPath="/opt/couchbase/samples/beer-sample.zip"
