+++ mh/medici/couchbase/scripts/run.sh
@@ -23,13 +23,12 @@ clusterBucketType="$8"
 clusterBucketName="$9"
 clusterReplicatCount=${10}
 
-#clusterBucketName="CloudifyCouchbase${instanceID}"
 
 # args:
 # $1 the error code of the last command (should be explicitly passed)
 # $2 the message to print in case of an error
 # 
-# an error message is printed and the script exists with the provided error code
+# an error message is printed and the script exits with the provided error code
 function error_exit {
 	echo "$2 : error code: $1"
 	exit ${1}
@@ -50,7 +49,6 @@ function performPostStart {
 
 	# When using public ip. Seems to have to wait a little longer
 	sleep 20s
-	
 	echo "Couchbase is now ready for action"
 
 	$1 cluster-init -c $clusterHost:8091 --cluster-init-username=$clusterAdmin --cluster-init-password=$clusterPassword --cluster-init-port=8091 --cluster-init-ramsize=$clusterRamSize -d	
@@ -65,6 +63,7 @@ function performPostStart {
 		echo "Memory bytes: ${memory_bytes}"
 		echo "Memory MB: ${memory_mb}"
 		echo "Memory Allocation: ${memory_allocation}"
+	
 	
 		$1 bucket-create -u $clusterAdmin -p $clusterPassword -c $clusterHost:8091 --bucket=$clusterBucketName --bucket-type=$clusterBucketType --bucket-ramsize=$memory_allocation --bucket-replica=$clusterReplicatCount
 		    
@@ -100,11 +99,6 @@ if  [ $postStartRequired == "true" ]; then
 	echo "After performPostStart"
 fi
 
-# Stop the node for XDCR setup
-# sudo /etc/init.d/couchbase-server stop
-
-# Pick new port
-# sudo /etc/init.d/couchbase-server start
 
 ps -ef | grep -i couchbase | grep -viE "grep|gsc|gsa|gigaspaces"
