--- mh/services/couchbase/couchbase-service.properties
@@ -1,45 +1,0 @@
-/*******************************************************************************
-* Copyright (c) 2012 GigaSpaces Technologies Ltd. All rights reserved
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*       http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*******************************************************************************/
-serviceName="couchbase"
-
-couchbasePort=8097
-couchbaseStatsPort=11209
-
-couchbaseUser="admin"
-couchbasePassword="q12345a"
-
-clusterRamSize=512
-
-clusterBucketName ="medici"
-clusterBucketType ="couchbase"
-
-clusterReplicatCount=1
-
-cbVersion="2.0.0"
-/* Taken from http://www.couchbase.com/download */
-couchbaseRpmLinux32="http://packages.couchbase.com/releases/${cbVersion}/couchbase-server-community_x86_${cbVersion}.rpm"
-couchbaseRpmLinux64="http://packages.couchbase.com/releases/${cbVersion}/couchbase-server-community_x86_64_${cbVersion}.rpm"
-
-couchbaseDebUbuntu32="http://packages.couchbase.com/releases/${cbVersion}/couchbase-server-community_x86_${cbVersion}.deb"
-couchbaseDebUbuntu64="http://packages.couchbase.com/releases/${cbVersion}/couchbase-server-community_x86_64_${cbVersion}.deb"
-
-couchbaseExeWindows32="http://packages.couchbase.com/releases/${cbVersion}/couchbase-server-community_x86_${cbVersion}.setup.exe"
-couchbaseExeWindows64="http://packages.couchbase.com/releases/${cbVersion}/couchbase-server-community_x86_64_${cbVersion}.setup.exe"
-
-couchbaseZipMac64="http://packages.couchbase.com/releases/${cbVersion}/couchbase-server-community_x86_64_${cbVersion}.zip"
-
-/* Only for a demo */
-dataPath="/opt/couchbase/samples/beer-sample.zip"
