--- mh/services/couchbase/scripts/stop.sh
@@ -1,17 +1,0 @@
-#!/bin/bash
-
-# args:
-# $1 the error code of the last command (should be explicitly passed)
-# $2 the message to print in case of an error
-# 
-# an error message is printed and the script exits with the provided error code
-function error_exit {
-	echo "$2 : error code: $1"
-	exit ${1}
-}
-
-export PATH=$PATH:/usr/sbin:/sbin:/opt/couchbase/bin || error_exit $? "Failed on: export PATH=$PATH:/usr/sbin:/sbin:/opt/couchbase/bin"
-
-
-sudo /etc/init.d/couchbase-server stop
-
